Changes to the overall look, especially icons
Added icons from fams silk iconset. They are free and give a consistent look.
The old icons had transparency issues.
Replaced the favicon with one created from silk set.
Changed the symbol to use as a bookmark link to text with another silk icon.
Mainly because of consistency.
Replaced plus and minus icons with text though it is not a final solution.
Removed the footer from 'normal' pages.
It doesn't need to be on every page. The z-index made it hover over important content.
It is still displayed on 'settings' and 'admin' page.